Add pganalyze-collector-helper for accessing restricted system info.
Unfortunately since we're running the collector as a non-priviledged user
we can't read all data that we need. Specifically we need to determine
the location of the data directories so we can correctly assign disk I/O
and usage.

This introduces a helper process that runs as root using setuid, but can
only be called by members of the pganalyze group. This is still better
than having the whole collector run as root, and the helper is much easier
to audit.