Introduce ability to collect statement text less often
For some systems it is significantly more expensive to query
pg_stat_statements together with the query text, since often the query
text file can grow to 100MB+ in size.

This introduces a server-controlled option that only collects the
statement text every N invocations, and remembers the statistics in all
other runs.