Allow enabling/disabling automatic EXPLAIN from server-side
This makes auto-explain behave the same way as logs. It defaults to off
on the server, on in the client - which means it doesn't run by default,
but can be enabled without changing something in the client. Nonetheless
the client settings can be used to override the server option by using
--no-explain.