Add experimental support for Docker log monitoring
This allows monitoring the logs of a Postgres instance running inside
Docker, when running the collector outside (on the Docker Host). This
does not support monitoring a Postgres container with the collector
running as a sidecar container inside Docker, yet.

Configure "db_log_docker_tail" on a server, and set it to the Docker
container name to collect its log output. The collector needs to have
sufficient permissions to run "docker logs" from its process.

This is experimental and primarily intended to be used for pganalyze
development and testing purposes.