Introduce log filtering for PII and other kinds of secrets
This introduces a new mechanism to filter sensitive data in the logs and
replace it with a character ("X" for now) instead of the actual data,
before sending it to the log storage.

The mechanism is controlled by the "filter_log_secret" configuration
setting, which defaults to "none" for now.

In addition this introduces the "filter_query_sample" config setting
which allows to remove all query samples that would otherwise be sent.