Fix helper process when using systemd service
This unfortunately requires us to remove the "NoNewPrivileges" mode, since
we intentionally use a setuid binary to be able to discover the data
directory as well as determine the size of the WAL directory.

We may refactor this in the future into a separate service to avoid this,
but for now this unbreaks the helper for self-hosted systems.