Increase systemd service memory limit to 1GB
We've previously limited this to 256MB for all use cases, which is too
small when monitoring multiple systems.

This limit exists to avoid runaway collector processes, in particular
when running the collector directly on the database server. On production
databases today, 1GB of memory doesn't come free, but is still a safe
amount to use in spikes.