Add support for "[%p]: [%l-1] db=%d,user=%u " log_line_prefix
This is intended for cases where the timestamp of the log line is
provided separately (like for the "simple" prefix), and should *not*
be used for other scenarios like self-managed servers.

In addition, this adds consistent parsing of the log line number. Whilst
not commonly used, its still a factor thats used in sorting the log
output, and there is no reason to not parse it when its there.