Split out activity and log snapshot states and mutexes
This improves consistency of when these snapshots run, because they
no longer conflict with each other, or the full snapshot run.

Note that for now, the state file does not contain the activity and log
states. This could be implemented in the future, as long as its ensured
that there is a write mutex when writing the state file, as this could
happen concurrently now.