High-frequency query collection: Avoid race condition, run in parallel
Previously this was first reading the server values, and then waiting on
the mutex. In the case a full snapshot took longer than 1 minute, that
meant that this was waiting for a full snapshot in the mutex.

Unfortunately, the full snapshot writes the server.PrevState value at the
very end of its run - which was discarded based on the fact that the
compact snapshot worker had already cached an older previous state in the
"server" variable.

Small as it may seem, this was the cause of a slow but steady memory leak,
because the UnidentifiedStatementStats field of the PersistedState struct
just kept growing.

This also may have caused incorrect values to show for other statistics,
due to diffs not using the correct previous values.

In passing, parallelize the collection of the high-frequency query
statistics, to improve accuracy when the collector is monitoring a lot
of servers.