Log processing: Allow multi-line log content to be processed
For current log providers, data always comes in once per-line, but for an
upcomig provider we need to accept multi-line logs as well. This does not
have any implications for currently supported log providers, that always
pass a single line to these regular expressions.

In passing, clean up a typo and remove an outdated comment.