Add ignore_schema_regexp (#62)
This is like ignore_table_pattern, but pushed down into the actual
stats-gathering queries to improve performance. This should work much
better on very large schemas.

We use a regular expression instead of the current glob-like matching
since the former is natively supported in Postgres.

Note that for vacuum activity, we filter stats like with everything else, but
since the fact that there is activity is interesting outside of the specific
tables being vacuumed, include that information but with a null table
reference.

We now warn on usage of the deprecated `ignore_table_pattern` field.