EXPLAIN plans / Query samples: Support log line prefix without %d and %u
Whilst not recommended, in some scenarios changing the log_line_prefix
is difficult, and we want to make it easy to get EXPLAIN data even in
those scenarios.

In case the log_line_prefix is missing the database (%d) and/or the user
(%u), we simply use the user and database of the collector connection.