Use more useful partioned_by field
Using the pg_get_partkeydef helper lets us use a field we can display
directly in the UI. With a direct pg_expr call on the partition
expressions, we would need to manually merge that with column
definitions. That's awkward because the field is an unstructured
string.

Since we're not using the field yet, we can just change the way we
populate the existing field rather than introduce a new field.