Try separate release workflow for GitHub actions
The push / pull_request actions don't actually trigger for tags; we
need to watch create for that. We could still theoretically share the
CI build, but there's no simple way to add workflow dependencies [1],
and we could be in a situation where a tag is pushed without a
corresponding branch push or pull request (e.g., when a tag for an
existing commit is pushed), so attempting that is dubious in the first
place.

[1]: There are projects using not-so-simple ways of doing it [2], but
     that does not seem worth it compared to occasionally running the
     build twice
[2]: https://github.com/asarkar/build-time-tracker/blob/master/.github/workflows/publish.yml#L19-L23