Log stream: Optimize stitching logic to allocate one large string
Previously we kept resizing the string, leading to many small allocations.
Since we have all the strings that we need already allocated, instead do
one new allocation once, for the final size of the string.

Note that there is further optimization potential here, but this aims to
reduce the biggest current problem point.