Log stream: Handle streams with missing PID/timestamps/etc correctly
Previously we would sort lines by PID, even when some of the lines do
not have a PID assigned, such as multi-line logs from when we process a
log tail from a self-managed server.

In those situations that lead to incorrect output. Instead, only consider
these characteristics when everything in the stream has the PID/timestamp
assigned.