Check and report problematic log collection settings (#95)
Some Postgres settings almost always cause a drastic increase in log
volume for little actual benefit. They tend to cause operational problems
for the collector (due to the load of additional log parsing) and the
pganalyze service itself (or indeed, likely for any service that would
process collector snapshots), and do not add any meaningful insights.
Furthermore, we found that these settings are often turned on
accidentally.

To avoid these issues, add some client-side checks in the collector to
disable log processing if any of the problematic settings are on.

The settings in question are:

 * log_min_duration_statement [1] less than 100ms
 * log_min_messages [2] set to 'all'
 * log_duration [3] set to 'on'

If any of these are set to these unsupported values, log collection will be
disabled. The settings are re-checked every full snapshot, and can be
reset with a collector reload.

[1]: https://www.postgresql.org/docs/current/runtime-config-logging.html#GUC-LOG-MIN-DURATION-STATEMENT
[2]: https://www.postgresql.org/docs/current/runtime-config-logging.html#GUC-LOG-MIN-MESSAGES
[3]: https://www.postgresql.org/docs/current/runtime-config-logging.html#GUC-LOG-DURATION