Introduce API system scope fallback
This fallback is intended to allow changing the API scope, either based
on user configuration (e.g. moving the collector between different
cloud provider accounts), or because of changes in the collector identify
system logic.

The new "api_system_scope_fallback" / PGA_API_SYSTEM_SCOPE_FALLBACK config
variable is intended to be set to the old value of the scope. When the
pganalyze backend receives a snapshot with a fallback scope set, and there
is no server created with the regular scope, it will first search the
servers with the fallback scope. If found, that server's scope will be
updated to the (new) regular scope. If not found, a new server will be
created with the regular scope. The main goal of the fallback scope is to
avoid creating a duplicate server when changing the scope value.