Use new fallback scope mechanism to change scope for RDS databases
Previously we identified RDS databases by there ID and region only, but
the ID does not have to be unique within a region, it only has to be
unique within the same AWS account in that region. Thus, adjust the
scope to include both the region and AWS Account ID (if configured or
auto-detected), and use the fallback scope mechanism to migrate existing
servers.