Handle NULL parameters for query samples correctly
Previously we assumed that the "parameters: " log line always contains
strings. Whilst that is generally correct, there is the exception of NULL
values, which get represented as "NULL" (without the quotes), instead of
having single quotes around them.

This would cause log-based EXPLAIN to break for these kind of samples,
as well as these values not showing up in-app.