Stop building packages for CentOS 6 and Ubuntu 14.04 (Trusty)
This is motivated by build failures due to package lists no longer working
for CentOS 6, and the expectation that something similar will happen for
Ubuntu 14.04 soon.

Both of these systems are now end of life, and the remaining survivor
of the CentOS 6 line (Amazon Linux 1) will be EOL on December 31st 2020.

This also means that all packages now built rely on systemd, and the
sysvinit/upstart packaging logic is removed.