Ensure we only set collected_local_catalog_data when appropriate
Right now, the flag indicates whether schema collection was successful
for a particular database, but we set it if we are at all able to
connect to a database, even if the actual data collection fails
afterward (we print warnings but that's not sufficient here).

Instead, only set the flag if collection of relation and function
metadata, and of relation and index stats, all succeed.