Validate db_url/DB_URL, if set, when initially reading config
Right now, we do not validate it, so if set to an invalid URL, it can
crash some of our helper functions with a segfault due to a nil
pointer reference.

Validate at startup, but allow to continue to avoid retrofitting error
handling across helpers that right now cannot return errors. The helpers
return dummy/default values which will fail authentication.