Change <unidentified queryid> marker to clarify cause
Right now, if pg_stat_statements.max is too low for a given query
workload, we may end up capturing stats for queries whose text is no
longer available to us. We flag this situation by using the string
"<unidentified queryid>" as the query text, but this is fairly opaque.
We have a change in progress to improve the display of this in-app,
but changing the actual marker string used here can help clarify
things as well.

Instead of "<unidentified queryid>", this patch changes the text for
queries like this to "<query text unavailable>", which describes the
situation more accurately.