Add support for receiving logs from remote servers over syslog
You can now specify the new "db_log_syslog_server" config setting, or
"LOG_SYSLOG_SERVER" environment variable in order to setup the collector
as a syslog server that can receive logs from a remote server via syslog
to the server that runs the collector.

Note that the format of this setting is "listen_address:port", and its
recommended to use a high port number to avoid running the collector as root.

For example, you can specify "0.0.0.0:32514" and then send syslog messages
to the collector's server address at port 32514.

Note that you need to use protocol RFC5424, with an unencrypted TCP
connection. Due to syslog not being an authenticated protocol it is
recommended to only use this integration over private networks.