Log parsing: Correctly handle %a containing commas/square brackets
Note that this does not support all cases since Go's regexp engine
does not support negative lookahead, so we can't handle an application
name containing a comma if the log_line_prefix has a comma following %a.

We may improve this further in the future by rewriting the log line
parser to not be regexp based.