Drop the log_directory helper command and just fetch the setting
The helper fails on a default AL2 install since the binary is called
postmaster not postgres, and there's no apparent reason not to use the
GUCS directly. This lets us drop the helper subcommand entirely.

Also, do not show a db_log_location if the log directory is inside the
data directory: this is almost certain to cause permissions problems,
and instead we can direct users to address this in our setup guide.