Update log location discovery (#149)
Drop the log_directory helper command and just fetch the setting.

The helper fails on a default AL2 install since the binary is called
postmaster not postgres, and there's no apparent reason not to use the
GUCS directly. This lets us drop the helper subcommand entirely.

Also, warn and only show relative location if log_directory is inside
the data directory: users will need to do some special configuration
to support this, so we should highlight it here and expand on it in
our setup guide.