Add integration test for install script
Since the script operates on published packages, it cannot test the
actual packages that would be built based on the tested commit, but it
can run the script from the tested commit with the latest published
packages. Since the install script has few interactions with the
package contents (except for assumptions about the overall structure),
this is still useful in testing the install script itself.