Docker builds: Increase stack size to 2MB to prevent rare crashes
Alpine has a very small stack size by default (80kb) which is less than
the default that Postgres expects (100kb). Since there is no good reason
to reduce it to such a small amount, increase to usually common Linux
default of 2MB stack size.

This would have surfaced as a hard crash of the Docker container with
error code 137 or 139, easily confused with out of memory errors, but
clearly distinct from it.