Also check for the name "postmaster" when looking for Postgres pid
Right now, the helper program we use for some lookups assumes the
Postgres process is named "postgres". This is true in most situations,
but on some systems (notably RHEL), the process is instead named
"postmaster", so we may fail to find it.

To work around this, look for "postgres" first but fall back to
"postmaster" if that fails.