Docker: Use Docker's USER command to set user, to support running as non-root
This enables the collector container to run in environments that require the
whole container to run as a non-root user, which previously was not the case.

For compatibility reasons the container can still be run as root explicitly,
in which case the setpriv command is used to drop privileges. setpriv replaces
gosu since its available for installation in most distributions directly, and
fulfills the same purpose here.