Statement stats retrieval: Get all rows first, before fingerprinting queries
This avoids showing a bogus ClientWrite event on the Postgres server side whilst
the collector is running the fingerprint method. There is a trade-off here,
because we now need to retrieve all statement texts (for the full snapshot) before
doing the fingerprint, leading to a slight increase in memory usage. Nonetheless,
this improves debuggability, and avoids bogus statement timeout issues.