Vacuum activity: Correctly handle output transform of duplicate tables
This previously used a simplistic model where we assumed we only ever
have one vacuum run referencing a certain table. However this is not
correct when the TOAST relation of a table is being vacuumed at the same
time as the table itself. In that case we'll have the same relation twice,
requiring the use of the more correct upsertRelationReference helper.

In passing, move the upsert helper functions into a combined util.go file.