Change schema collection message to warning
If we fail to collect schema metadata, a database can remain
permanently hidden and a lot of in-app functionality does not
work. Since verbose logs are not likely to be turned on for a running
collector, hiding this message makes this situation difficult to
discover.

Elevate the message to a warning.