Log Insights: Improve regexp for "malformed array literal" log event (#219)
Previously we did not support a double quote showing up inside the array
content, and would instead detect the rest of the log line as unknown
data. This adjusts the regexp to correctly capture that data and mark
it as a table data log secret.

Additionally, adds the known DETAIL line "Unexpected array element.".