Table processing: Ensure we use -1 index for unknown parent/FK tables
Previously we would sometimes use 0 in these situations, which is a valid
index when referencing the very first table in the list.

Note that whilst its not always clear why bad references here happen, as
they should only happen when the schema_table_regexp logic is active,
its better to not send bogus data to the server.