Log processing: Handle non-UTC/non-local log_timezone values correctly
Previously we supported handling timezone values in three cases:

(1) No timezone specified
(2) UTC/GMT timezone explicitly specified
(3) Timezone specified that matches the collector system timezone

However we did not handle a log_timezone setting that is different from
the timezone thats configured on the system the collector is running on.

This was caused by incorrect assumptions of what time.Parse() does in
this case. As we have no other way to get the UTC timestamp of a logline,
we rely on the system's timezone database to help us parse the log line
in the correct timezone for this edge case.