Normalization: Attempt auto-fixing truncated queries
This adjusts the normalization logic to use the same "fix unparsable
queries" logic that already existed for fingerprinting, and is mainly
intended for data in pg_stat_activity (that is often truncated).

This logic provides a better level of grouping for these truncated
queries.

Add the new "/* truncated-query */ " magic comment to make it clear to the
viewer that they are looking at a fixed truncated query text, and not
the full query text as sent by the application.