Handle booleans in config more consistently
Right now, we parse boolean values in our config files with go-ini,
but environment-based config goes through some ad-hoc parsing. For
most boolean values, any value other than the empty string and '0' is
considered true (for the DB_ALL_NAMES setting, any value except '1' is
considered false).

Unify this to treat case-insensitive false, off, no, 'f', and 'n' as
false in addition to zero. All other values are still considered true
to preserve backward compatibility (this breaks backward compatibility
for DB_ALL_NAMES, but that setting is unlikely to be set when
disabling it, and the unified behavior is worth the change).

This is closer to what go-ini does for parsing our config files [1],
and to the Postgres config file handling [2].

Fixes #245

[1]: https://ini.unknwon.io/docs/howto/work_with_values
[2]: https://www.postgresql.org/docs/current/config-setting.html