Log streams: Support processing primary and secondary lines out of order (#239)
This is in particular useful for GCP, where Pub/Sub does not guarantee
that associated lines will be received in order, e.g. when the DETAIL
line comes before the LOG line for log_min_duration_statement output,
causing parameters to be missed.

In passing, fix a bug in logline readiness handling that likely caused
some problems with Heroku logs, as we were considering logs to be "ready"
too often, due to a buggy comparison of the "logLine" loop variable
which we kept a long-lived pointer too (and then compared the next
iteration against that pointer, which would always succeed).