Log Insights: C22 Auth failed - detect additional DETAIL information
This avoids flagging the whole DETAIL log line as an unidentified line in
cases where the logdetail value is set by Postgres, e.g. because the
role that tried logging in does not exist (which is included in the
error message).

In passing, mark pg_hba.conf contents output by this log event as
OpsLogSecret.