Filter out vacuum records we cannot match to a table name
Currently we ensure that we have a v.relid, but that does not
guarantee that we match an entry in pg_class or in our
pg_stat_activity query.

If we have no table name to report (e.g., because of a manual VACUUM
running in a different database), omit the record from the vacuum
progress output.