Log processing: Add support for Azure DB for Postgres - Flexible Server
There are some differences in how Flexible Server sends data into the
Azure Event Hub as compared to Single Server, causing log lines to not
be processed. This attempts to parse lines for Flexible Server first,
and automatically falls back to the old special handling for
Single Server if that fails.