Log test: Don't count Heroku Free Tier as hard failure (different prefix)
We can't support log monitoring for Heroku Postgres' free tier, since
the log stream is not exported (intentionally, by Heroku), but previously
we would fail when running the test function, which may not be desirable
when multiple databases are mounted to the same collector application.

Instead emit a warning when the special log_line_prefix used by the
free tier is detected.