Query samples: Keep unmarshaled EXPLAIN JSON until output stage
This allows easier modification of the EXPLAIN before it gets unmarshaled
again in a future commit.

Note that we need to unmarshal the EXPLAIN JSON because its the only way
to get to the full query text (which we need to associate the log line
and query sample to the correct query).

In passing add missing fields for the top-level object (such as JIT and
Triggers), and use json.RawMessage for the top plan node to avoid
unnecessary unmarshaling.