Run post-processing of query samples in runner logic instead of output
It makes more sense logically, since we already have a function that
talks about post-processing in the runner logic.

Note that log filtering (i.e. masking secrets with "X") still has to
happen in the output logic, to avoid unnecessary loading of the log text
into memory, but we move the setting of the FilterLogSecret that controls
this into runner as well.