Add optional normalization of sensitive fields in EXPLAIN plans
This adds the new "normalize" option for the "filter_query_sample"
configuration setting. When this option is selected, EXPLAIN plans that
are collected from either auto_explain or log-based EXPLAIN are analyzed
on a field-by-field basis. Each field known to contain Postgres expressions
(which may contain constant values, which could contain PII), such as
"Filter" or "Index Cond", is ran through the same normalization function
applied to query texts. Unknown EXPLAIN fields are discarded when this
option is active.

Turning on this setting will also cause all query samples (whether they
have an EXPLAIN attached or not) to have their query text normalized
and their parameters marked as "<removed>".

This setting is recommended when EXPLAIN plans may contain sensitive
data that should not be stored. Please verify that the logic works
as expected with your workload and log output.