Add support for Aiven system type (#287)
Treat the Aiven Project ID as system scope and the Service ID as a
system ID. Aiven Postgres URLs include the service id separated from
the project id by a dash. The service id includes a dash, and the project
id may include a dash, but we can assume the service id has a standard
format (pg-139acf), so we take everything after the second dash to be the
project ID, and everything before the second dash to be the service ID.

Also add Fallback handling for SystemID and SystemType. This will help
us maintain identify for Aiven systems that were previously recognized as
self-managed.

Also track Aiven Project ID and Service ID in collector config metadata for
debugging.