Allow using "pg_read_file" for self-managed servers to read log files
This is an alternate approach to collecting log data that relies on
built-in Postgres functions to read log files and return the log data to
the collector over the Postgres connection.

Note that this requires superuser (either directly or through a helper)
and thus does not work on managed database providers, with the exception
of Crunchy Bridge, for which this is already the mechanism to fetch logs.

Additionally, this carries higher overhead than directly tailing log
files, or using syslog, and thus should only be used when necessary.

You can enable this by setting the "db_log_pg_read_file" /
"LOG_PG_READ_FILE" setting to "1" or "true", and ensuring that the
"pganalyze.read_log_file" helper exists (or by connecting as a superuser).