Show RDS instance role hint when running collector test (#306)
In fd65610fd60b0aaa2b2784ba23bb7f7e09c7f71e, we added a hint to
instruct users to check the instance role when hitting RDS credential
erros in log download. This works, but only in the live code path,
because the test code logs the download error from a different entry
point.

To work around this, add the same warning to the test code path.