RDS logs: Don't silence errors by accident due to variable shadowing
The way the existing code was written was creating a new local
variable called "err" that shadowed the top-level "err" variable. This
then lead to errors being silenced since the ErrorCleanup code path
was referencing the top-level "err" which was set to nil even in error
conditions.

The practical consequence of this was that if an IAM policy was in use
that permitted DescribeDBLogFiles but not DownloadDBLogFilePortion, the
log data would be empty, without any errors shown in the logs. This
can occur when someone grants the RDS Read Only predefined IAM policy
instead of a custom policy with our recommended permissions.