GH actions: Only apply permissions restrictions to build job (#314)
This was added in 8a5387cefb1287d0ed81a3b4a40dab811b2f6e45 to avoid the protoc
action using a more privileged token to do GH downloads, but broke the release
step by accident. For the release action only apply the permission limits to
the build step. For the CI action we keep the limits for the whole action.