Debian builds: Pin public_suffix package to version that supports old Ruby
For now we are still building with a very old Ruby here, as part of
building on an old Debian that generates a low minimum glibc version.

This unfortunately causes issues with gems that (understandably) drop
support for very old Ruby versions. This was the case previously with
fpm, but now also affects a dependency of fpm, public_suffix. In line
with fpm, pin public_suffix to the last release that supports our Ruby
version.