Statements: Fix Postgres 15 compatibility due to version check bug
There was a bug in our version check mechanism that caused the minor
version "1.10" to be treated as smaller than "1.8", effectively causing
us to look for the wrong fields.

Fix the relevant version comparisons to use integer instead of floats,
and in passing cleanup some old references to Postgres version numbers
as compared to pg_stat_statements version numbers.

In passing add the version strings for Postgres 14 and 15, in case we
add version-specific checks in the future.