RPM packages: Add RHEL9, Rocky Linux 8 and 9, and Fedora 36
For RHEL-based distributions that directly follow the RHEL release cycle
(e.g. Rocky Linux) the packages are located in the same namespace ("el/9")
but we explicitly test Rocky Linux going forward, as we've tested CentOS
before.

In passing this switches the tests for RHEL itself to use the Docker Hub
images now published for the Universal Base Image (UBI) for consistency
with other distros, instead of relying on images hosted within Red Hat's
Docker registry.