Collect Citus distributed index sizes
Currently, when monitoring Citus databases, we collect distributed
table sizes with Citus' `citus_table_size` helper function, but we
ignore indexes on these distributed tables.

Add a similar calculation for index sizes. While there is a similar
`citus_relation_size` that is documented to work for distributed
indexes, it does not [1]. Instead, we use `run_command_on_shards` to
get this information from the underlying catalog objects and stats
directly.

[1]: I filed https://github.com/citusdata/citus/issues/6496 which
     seems to be getting some attention, so maybe there is hope for a
     better fix